+++ PlanBuild/Blueprints/BlueprintAssets.cs
@@ -6,7 +6,6 @@ using PlanBuild.Utils;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
-using TMPro;
 using Object = UnityEngine.Object;
 
 namespace PlanBuild.Blueprints