+++ PlanBuild/ModCompat/PatcherAzuCraftyBoxes.cs
@@ -1,6 +1,5 @@
 using HarmonyLib;
 using PlanBuild.Plans;
-using System;
 using System.Collections.Generic;
 using UnityEngine;