+++ PlanBuild/ModCompat/PatcherCraftFromContainers.cs
@@ -1,19 +1,26 @@
 ﻿using HarmonyLib;
 using PlanBuild.Plans;
-using System;
 using System.Collections.Generic;
+
+using CraftyContainers = CraftFromContainers.CraftFromContainers;
 
 namespace PlanBuild.ModCompat
 {
     internal class PatcherCraftFromContainers
     {
-        [HarmonyPatch(typeof(PlanPiece), "GetInventories")]
+        [HarmonyPatch(typeof(PlanPiece), nameof(PlanPiece.GetInventories))]
         [HarmonyPostfix]
-        private static void PlanPiece_GetInventories_Postfix(Humanoid player, ref List<PlanPiece.IInventory> __result)
+        private static void PlanPiece_GetInventories_Postfix(
+            Humanoid player,
+            ref List<PlanPiece.IInventory> __result
+        )
         {
             // Check if the plugin is active, and append any nearby container inventories to the list
-            if (CraftFromContainers.BepInExPlugin.modEnabled.Value) {
-                foreach (Container container in CraftFromContainers.BepInExPlugin.GetNearbyContainers(player.transform.position)) {
+            if (CraftyContainers.modEnabled.Value) {
+                List<Container> containers = CraftyContainers.GetNearbyContainers(
+                    player.transform.position
+                );
+                foreach (Container container in containers) {
                     __result.Add(new PlanPiece.StandardInventory(container.GetInventory()));
                 }
             }